Disable select lowering
G_SELECT instructions really should be scheduled along with everything
else. It's probably fine to lower libcalls in legalizer, since we're
going to build a scheduling algorithm that can schedule around calls.
But select lowering really dices the original CFG to bits, and it makes
it very hard to see the original logic. Selects are more like real
instructions anyway; we may be able to treat them directly in the
scheduler.